fix: Fix crash if there is no newline at the end of the last line
The `Range` may extend the actual range of the text according to the doc
of LSP:
> ... the end position is exclusive. If you want to specify a range
> that contains a line including the line ending character(s) then use
> an end position denoting the start of the next line.

If there is no newline at the end of the last line (which is common for
files created by Vistual Studio) and client set the end of Range to the
start of the next line of the last line, then csharp-ls will crash
because argument is out of the range. On the other hand, we SHOULDN'T
trust the value client sent to us.

Signed-off-by: Adam Tao <tcx4c70@gmail.com>