feat(documentSymbol): Enrich Name of DocumentSymbol
ISymbol.Name will only show the name of the symbol, which contains
little information for method/function, and it will show ".ctor" for
constructor.

Therefore, use SymbolDisplayFormat to configure the information showed
in Name of DocumentSymbol. It will only show name for fields/properties,
and will show parameter types & names for method/function.

Signed-off-by: Adam Tao <tcx4c70@gmail.com>