refactor(RoslynHelpers.fs): Pass TextSpan to DocumentSymbolCollector.push
The identifier parameter of push is only used to provide position
information of the token, which can be done by pass a TextSpan to it.

For some SyntaxNodes (e.g., ConversionOperatorDeclarationSyntax), it's
not easy to get the SyntaxToken for the identifier while it's very easy
to get the TextSpan for the identifier. Therefore, change to pass
TextSpan to it which will make further changes much easier.

Signed-off-by: Adam Tao <tcx4c70@gmail.com>