Use a common target framework from all of your projects
Summary:

When using multiple target frameworks in multiple projects. We are currently
resolving the latest target framework that exists over all of your projects. The
issue with this is the selected target framework may not work with some of the
projects.

This resolves the latest target framework that exists in all of your projects.
This will give the best framework compatibility.

Test Plan:

Unit test for the existing select latest framework is still there. I have also
tested this locally over a project that uses `net8.0` and `net10.0` in some
projects.

Ref: #75